Changing the name of forms classes
We think the name that we put in forms elements was very confusing. The
label was called .ls-field and the text of label was called ls-field.
Now, we chage this to this:
label - ls-label
text of label - ls-label-text
input/field - ls-field
fieldset or another form container - ls-form-group

We think this is better.