Creating a trigger status
The problem:
When the sidebar is minimised, the submenu can’t stay opened. But, if
it have a child active, when the submenu is maximized, the submenu need
to be opened.

I create a trigger to know when the sidebar was maximized or minimized.
The function hasSubmenuItemActive return me if the submenu have some
child is .ls-active
In openSubmenuItemActive, I hear the hasSubmenuItemActive return. If is
true, I add a class ls-active in ls-submenu-parent.

When the sidebar is toggled, I need to close the Submenu.
In function closeSubmenu, I check if the sidebar toggle trigger is
fired, if yes, I check if ls-submenu-parent is activated, if yes, I
remove the class ls-active.