Changing the way we can use the classes to increase modal size. Closes #1700
Before, we need to use two classes to change the modal size:
`ls-modal-box and ls-modal-large` to make big modal, ou `ls-modal-box
ls-modal-small` to make small modal. Now, you can use only
`ls-modal-large` or `ls-modal-small`, and the old way works too.

TODO: we need to change where we put this classes. The right place is
add a classe ls-modal-large|small in the element ls-modal, because this
element is the parent of modal.

This Closes #1700